cart handling in progress

+++ src/main/webapp/assets/Scripts/cartHandler.js
@@ -1,35 +1,62 @@
 var cartCount = 0;
-function addToCartWhenLogin(id, quantity) {
-  var cartMessage = document.getElementById("cartMessage");
-  console.log(id);
+function addToCartWhenLogin(
+  id,
+  productName,
+  productPrice,
+  productDescription,
+  productImage,
+  isLogin
+) {
+  var product = {
+    productId: id,
+    name: productName,
+    price: productPrice,
+    description: productDescription,
+    image: productImage,
+    quantity: 1,
+  };
 
+  console.log(product);
   // Retrieve existing cart items from local storage or initialize an empty array
+  var nameProduct = $("#addToCartButton")
+    .closest(".product-container")
+    .find(".js-name-detail")
+    .text();
+  swal(nameProduct, "is added to cart !", "success");
   var cartItems = JSON.parse(localStorage.getItem("cartItems")) || [];
+  var existingProductIndex = cartItems.findIndex(function (p) {
+    return p.productId === id;
+  });
+  //   // If the product already exists, update its quantity
 
-  // Add the new product ID to the cart items array if it doesn't already exist
-  if (!cartItems.includes(id)) {
-    cartItems.push(id);
+  if (existingProductIndex !== -1) {
+    console.log("terminate because product already exist");
+    return;
   }
+  cartItems.push(product);
 
   // Save the updated cart items array back to local storage
   localStorage.setItem("cartItems", JSON.stringify(cartItems));
-
-  addCartItemToDB(id);
+  loadProductsToCart(); // thios method from side-cart.jsp
+  if (isLogin == true) addCartItemToDB(id);
 }
 
 function addCartItemToDB(id) {
   var data = new URLSearchParams();
   data.append("productId", id);
-
+  console.log("isnside db method js");
   fetch("front?controller=addToCart", {
     method: "POST",
     body: data,
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded",
+    },
   })
     .then((response) => response.text())
     .then((data) => {
-      console.log(data);
       if (data == "false") {
         console.log("out of stock");
+        var cartMessage = document.getElementById("cartMessage");
         cartMessage.innerText = "Sorry it is out of stock";
       } else {
         cartCount++;
